pack model to lightning
