--- data/flowers/tiny_splits.csv
@@ -1,30 +1,0 @@
-,Unnamed: 0,split,img_path,sem_path,inst_path
-0,0,train,data/flowers/A1/plant001_rgb.png,data/flowers/A1/plant001_fg.png,data/flowers/A1/plant001_label.png
-1,1,train,data/flowers/A1/plant005_rgb.png,data/flowers/A1/plant005_fg.png,data/flowers/A1/plant005_label.png
-2,2,train,data/flowers/A1/plant006_rgb.png,data/flowers/A1/plant006_fg.png,data/flowers/A1/plant006_label.png
-3,3,train,data/flowers/A1/plant007_rgb.png,data/flowers/A1/plant007_fg.png,data/flowers/A1/plant007_label.png
-4,4,train,data/flowers/A1/plant008_rgb.png,data/flowers/A1/plant008_fg.png,data/flowers/A1/plant008_label.png
-115,115,train,data/flowers/A2/plant004_rgb.png,data/flowers/A2/plant004_fg.png,data/flowers/A2/plant004_label.png
-116,116,train,data/flowers/A2/plant006_rgb.png,data/flowers/A2/plant006_fg.png,data/flowers/A2/plant006_label.png
-117,117,train,data/flowers/A2/plant007_rgb.png,data/flowers/A2/plant007_fg.png,data/flowers/A2/plant007_label.png
-118,118,train,data/flowers/A2/plant008_rgb.png,data/flowers/A2/plant008_fg.png,data/flowers/A2/plant008_label.png
-145,145,train,data/flowers/A3/plant008_rgb.png,data/flowers/A3/plant008_fg.png,data/flowers/A3/plant008_label.png
-146,146,train,data/flowers/A3/plant009_rgb.png,data/flowers/A3/plant009_fg.png,data/flowers/A3/plant009_label.png
-147,147,train,data/flowers/A3/plant010_rgb.png,data/flowers/A3/plant010_fg.png,data/flowers/A3/plant010_label.png
-148,148,train,data/flowers/A3/plant011_rgb.png,data/flowers/A3/plant011_fg.png,data/flowers/A3/plant011_label.png
-244,244,train,data/flowers/A4/plant0091_rgb.png,data/flowers/A4/plant0091_fg.png,data/flowers/A4/plant0091_label.png
-245,245,train,data/flowers/A4/plant0092_rgb.png,data/flowers/A4/plant0092_fg.png,data/flowers/A4/plant0092_label.png
-246,246,train,data/flowers/A4/plant0093_rgb.png,data/flowers/A4/plant0093_fg.png,data/flowers/A4/plant0093_label.png
-247,247,train,data/flowers/A4/plant0095_rgb.png,data/flowers/A4/plant0095_fg.png,data/flowers/A4/plant0095_label.png
-712,712,dev,data/flowers/A1/plant015_rgb.png,data/flowers/A1/plant015_fg.png,data/flowers/A1/plant015_label.png
-717,717,dev,data/flowers/A2/plant013_rgb.png,data/flowers/A2/plant013_fg.png,data/flowers/A2/plant013_label.png
-718,718,dev,data/flowers/A3/plant005_rgb.png,data/flowers/A3/plant005_fg.png,data/flowers/A3/plant005_label.png
-719,719,dev,data/flowers/A4/plant0817_rgb.png,data/flowers/A4/plant0817_fg.png,data/flowers/A4/plant0817_label.png
-747,747,test,data/flowers/A1/plant141_rgb.png,data/flowers/A1/plant141_fg.png,data/flowers/A1/plant141_label.png
-748,748,test,data/flowers/A1/plant142_rgb.png,data/flowers/A1/plant142_fg.png,data/flowers/A1/plant142_label.png
-757,757,test,data/flowers/A2/plant002_rgb.png,data/flowers/A2/plant002_fg.png,data/flowers/A2/plant002_label.png
-758,758,test,data/flowers/A2/plant031_rgb.png,data/flowers/A2/plant031_fg.png,data/flowers/A2/plant031_label.png
-759,759,test,data/flowers/A3/plant001_rgb.png,data/flowers/A3/plant001_fg.png,data/flowers/A3/plant001_label.png
-760,760,test,data/flowers/A3/plant006_rgb.png,data/flowers/A3/plant006_fg.png,data/flowers/A3/plant006_label.png
-761,761,test,data/flowers/A4/plant0258_rgb.png,data/flowers/A4/plant0258_fg.png,data/flowers/A4/plant0258_label.png
-762,762,test,data/flowers/A4/plant0428_rgb.png,data/flowers/A4/plant0428_fg.png,data/flowers/A4/plant0428_label.png
