+++ pyproject.toml
@@ -17,6 +17,7 @@ seaborn = "^0.13.0"
 pandas = "^2.1.1"
 omegaconf = "^2.3.0"
 albumentations = "^1.3.1"
+lightning = "^2.1.0"
 
 [tool.poetry.group.dev.dependencies]
 pre-commit = "^3.4.0"
